+++ plugins/db.mysql.editors/macosx/MySQLTableEditor.mm
@@ -434,7 +434,7 @@ static NSString *columnDragUTI = @"com.mysql.workbench.column";
 {
   [mColumnsTable reloadData];
   
-  int rowIndex = [mColumnsTable selectedRow];
+  NSInteger rowIndex = [mColumnsTable selectedRow];
   if (rowIndex >= 0) {
     NSString* collation = [mColumnsDataSource objectValueForValueIndex: bec::TableColumnsListBE::CharsetCollation
                                                                    row: rowIndex];
@@ -511,7 +511,7 @@ static NSString *columnDragUTI = @"com.mysql.workbench.column";
 
   [mIndicesTable reloadData];
 
-  int rowIndex = [mIndicesTable selectedRow];
+  NSInteger rowIndex = [mIndicesTable selectedRow];
   if (rowIndex >= 0) {
     NSString* indexName = [mIndicesDataSource objectValueForValueIndex: MySQLTableIndexListBE::Name
                                                                    row: rowIndex];
@@ -559,7 +559,7 @@ static NSString *columnDragUTI = @"com.mysql.workbench.column";
       MFillPopupButtonWithStrings((NSPopUpButton*)cell, mBackEnd->get_fks()->get_columns()->get_ref_columns_list(rowIndex, false));
   }
   
-  int rowIndex = [mFKTable selectedRow];
+  NSInteger rowIndex = [mFKTable selectedRow];
   
   if (rowIndex >= 0) {
     NSString* fkName = [mFKDataSource objectValueForValueIndex: bec::FKConstraintListBE::Name
@@ -568,7 +568,7 @@ static NSString *columnDragUTI = @"com.mysql.workbench.column";
     
     mBackEnd->get_fks()->select_fk(rowIndex);
     
-    int rowIndex = [mFKTable selectedRow];
+    NSInteger rowIndex = [mFKTable selectedRow];
     NSString* updateAction = [mFKDataSource objectValueForValueIndex: bec::FKConstraintListBE::OnUpdate
                                                                  row: rowIndex];
     [mFKOnUpdate selectItemWithTitle: updateAction];
@@ -1136,13 +1136,13 @@ objectValueForTableColumn: (NSTableColumn*) aTableColumn
       if (lastRowIndex == rowIndex)
         [self activateColumnPlaceholder: rowIndex];
       [mColumnsDataSource setIntValue: [value intValue]
-                        forValueIndex: valueIndex
+                        forValueIndex: (int)valueIndex
                                   row: rowIndex];
     }
     else {
       if (keepEditingTypeString == nil) {        
         [mColumnsDataSource setStringValue: value
-                             forValueIndex: valueIndex
+                             forValueIndex: (int)valueIndex
                                        row: rowIndex];
       }
       else {
@@ -1465,7 +1465,7 @@ objectValueForItemAtIndex: (NSInteger) index
     [self updateFKPlaceholder];
   }
   else if( sender == mColumnsCollation) {
-    int rowIndex = [mColumnsTable selectedRow];
+    NSInteger rowIndex = [mColumnsTable selectedRow];
     if ([popItemTitle isEqualToString: @"Table Default"])
       popItemTitle = @" - ";
     [mColumnsDataSource setStringValue: popItemTitle
@@ -1473,7 +1473,7 @@ objectValueForItemAtIndex: (NSInteger) index
                                    row: rowIndex];
   }
   else if (sender == mIndicesStorageTypes) {
-    int rowIndex = [mIndicesTable selectedRow];
+    NSInteger rowIndex = [mIndicesTable selectedRow];
     NSString* storageType = popItemTitle;
     [mIndicesDataSource setStringValue: storageType
                          forValueIndex: MySQLTableIndexListBE::StorageType
@@ -1492,7 +1492,7 @@ objectValueForItemAtIndex: (NSInteger) index
   }
   
   else if (sender == mFKOnUpdate) {
-    int rowIndex = [mFKTable selectedRow];
+    NSInteger rowIndex = [mFKTable selectedRow];
     if (![mFKDataSource setStringValue: popItemTitle
                     forValueIndex: bec::FKConstraintListBE::OnUpdate
                               row: rowIndex])
@@ -1504,7 +1504,7 @@ objectValueForItemAtIndex: (NSInteger) index
     }
   }
   else if (sender == mFKOnDelete) {
-    int rowIndex = [mFKTable selectedRow];
+    NSInteger rowIndex = [mFKTable selectedRow];
     if (![mFKDataSource setStringValue: popItemTitle
                     forValueIndex: bec::FKConstraintListBE::OnDelete
                               row: rowIndex])
@@ -1687,14 +1687,14 @@ objectValueForItemAtIndex: (NSInteger) index
     [self updateTitle: [self title]];
   }
   else if (sender == mIndicesBlockSize) {
-    int rowIndex = [mIndicesTable selectedRow];
+    NSInteger rowIndex = [mIndicesTable selectedRow];
     NSString* blockSize = [mIndicesBlockSize stringValue];
     [mIndicesDataSource setStringValue: blockSize
                          forValueIndex: MySQLTableIndexListBE::RowBlockSize
                               row: rowIndex];
   }
   else if (sender == mIndicesParser) {
-    int rowIndex = [mIndicesTable selectedRow];
+    NSInteger rowIndex = [mIndicesTable selectedRow];
     NSString* blockSize = [mIndicesParser stringValue];
     [mIndicesDataSource setStringValue: blockSize
                          forValueIndex: MySQLTableIndexListBE::Parser
@@ -1787,21 +1787,21 @@ objectValueForItemAtIndex: (NSInteger) index
   }
   else if (sender == mColumnsComment || sender == mColumnsComment2) {
     [[aNotification object] breakUndoCoalescing];
-    int rowIndex = [mColumnsTable selectedRow];
+    NSInteger rowIndex = [mColumnsTable selectedRow];
     [mColumnsDataSource setStringValue: [sender string]
                          forValueIndex: bec::TableColumnsListBE::Comment
                               row: rowIndex];
   }
   else if (sender == mIndicesComment) {
     [[aNotification object] breakUndoCoalescing];
-    int rowIndex = [mIndicesTable selectedRow];
+    NSInteger rowIndex = [mIndicesTable selectedRow];
     [mIndicesDataSource setStringValue: [mIndicesComment string]
                          forValueIndex: bec::IndexListBE::Comment
                                    row: rowIndex];
   }
   else if (sender == mFKComment) {
     [[aNotification object] breakUndoCoalescing];
-    int rowIndex = [mFKTable selectedRow];
+    NSInteger rowIndex = [mFKTable selectedRow];
     [mFKDataSource setStringValue: [mFKComment string]
                     forValueIndex: bec::FKConstraintListBE::Comment
                          row: rowIndex];
